feature(gulp-bump): allow configuration after pipe construction
Because passed in options are bound to individual variables, they can not be changed after gulp-bump() has been add to a pipe. By removing these assignments an object can be passed to burp which could be altered in other steps of the pipeline. This would allow the following example to function:

```
var bumpOpts = {};
return gulp.src(['bower.json', 'package.json'])
      .pipe(require('gulp-prompt').prompt({
          type: 'list',
          name: 'type',
          message: 'What type of bump would you like to do for this release?',
          choices: ['major', 'minor', 'patch', 'prerelease'],
          default: 2
      }, function (res) {
          bumpOpts.type = res.type;
      }))
      .pipe(bump(bumpOpts))
      .pipe(gulp.dest('.'));
```