fix(gulp-bump): changed one too many variables
On line 74 a variable `value` was introduced that shadowed the old `value` variable. I changed this to use `opts.value` which is incorrect.

This means two things:
If you were to reject this pull request, variable shadowing should be avoided and this should probably be renamed.
If you accept this pull request (I hope you do), variable shadowing is no longer an issue.